fix typo

+++ train.py
@@ -156,7 +156,7 @@ def main():
                 while count_train < train_data_size:
                     nextcount = min(count_train + BATCH_SIZE, train_data_size)
                     x_train, y_train = load_dataset(res_dir, count_train, nextcount)
-                    err_train += sess.run(loss_d(y, y_pred, D)*D/train_data_size, feed_dict={x: x_train, y: y_train, D: nextcount - count_train, keep_prob: 1.})
+                    err_train += sess.run(loss_d(y, y_pred, D)*tf.cast(D, tf.float32)/train_data_size, feed_dict={x: x_train, y: y_train, D: nextcount - count_train, keep_prob: 1.})
                     count_train = nextcount
 
                 count_test = 0
@@ -164,7 +164,7 @@ def main():
                 while count_test < test_data_size:
                     nextcount = min(count_test + BATCH_SIZE, test_data_size)
                     x_test, y_test = load_dataset(res_dir, count_train + count_test, count_train + nextcount)
-                    err_test += sess.run(loss_d(y, y_pred, D)*D/test_data_size, feed_dict={x: x_test, y: y_test, D: nextcount - count_test, keep_prob: 1.})
+                    err_test += sess.run(loss_d(y, y_pred, D)*tf.cast(D, tf.float32)/test_data_size, feed_dict={x: x_test, y: y_test, D: nextcount - count_test, keep_prob: 1.})
                     count_test = nextcount
 
                 err_w = sess.run(loss_w())
